+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/AppContext.kt
@@ -64,6 +64,8 @@ class AppContext(
   // We postpone creating the `JSIInteropModuleRegistry` to not load so files in unit tests.
   internal lateinit var jsiInterop: JSIInteropModuleRegistry
 
+  private var hostWasDestroyed = false
+
   /**
    * The core module that defines the `expo` object in the global scope of the JS runtime.
    *
@@ -132,9 +134,7 @@ class AppContext(
   }
 
   fun onCreate() = trace("AppContext.onCreate") {
-    registry.readyForPostingEvents()
-    registry.post(EventName.MODULE_CREATE)
-    registry.flushTheEventQueue()
+    registry.postOnCreate()
   }
 
   /**
@@ -314,6 +314,12 @@ class AppContext(
       "Current Activity is of incorrect class, expected AppCompatActivity, received ${currentActivity?.localClassName}"
     }
 
+    // We need to re-register activity contracts when reusing AppContext with new Activity after host destruction.
+    if (hostWasDestroyed) {
+      hostWasDestroyed = false
+      registry.registerActivityContracts()
+    }
+
     activityResultsManager.onHostResume(activity)
     registry.post(EventName.ACTIVITY_ENTERS_FOREGROUND)
   }
@@ -331,6 +337,10 @@ class AppContext(
       activityResultsManager.onHostDestroy(it)
     }
     registry.post(EventName.ACTIVITY_DESTROYS)
+
+    // The host (Activity) was destroyed, but it doesn't mean that modules will be destroyed too.
+    // So we save that information, and we will re-register activity contracts when the host will be resumed with new Activity.
+    hostWasDestroyed = true
   }
 
   internal fun onActivityResult(activity: Activity, requestCode: Int, resultCode: Int, data: Intent?) {